Finalizando alterações no modulo runner

+++ src/main/java/block/DataSplit.java
@@ -3,7 +3,7 @@ package block;
 import java.io.IOException;
 
 public abstract class DataSplit {
-    public abstract void run(Double unitTime,Double initTime);
+    public abstract void run(Long unitTime,Long initTime);
     public abstract void insertData(String path);
     public abstract int getNumberPartitions();
 }
