+++ src/main/java/block/DataSplitFactory.java
@@ -3,7 +3,7 @@ package block;
 public abstract class DataSplitFactory {
     public abstract void createDataSplit(String logs, String data);
 
-    public abstract void run(Double unitTime, Double initTime);
+    public abstract void run(Long unitTime, Long initTime);
 
     public abstract int getNumberPartitions();
 }
